Add of the audio tour tool jar file
This is the jar file for the audio tour tool.  The tool let's you highlight lines of code and record audio for the highlighted lines and save them in a file with the same prefix.  The rst for the activecode with an audio tour can be generated by clicking the generate button.